+++ bigchem/__init__.py
@@ -1,0 +1,3 @@
+from .algos import *
+from .canvas import *
+from .tasks import *