+++ examples/terachem.py
@@ -1,54 +1,59 @@
-"""How to perform a basic, one-task calculation using BigChem"""
+"""How to perform a basic, single program calculation using BigChem"""
+from pathlib import Path
 
-from pprint import pprint
+from qcio import CalcType, Molecule, ProgramInput
 
-import qcengine as qcng
-from qcelemental.models import AtomicInput
+from bigchem import compute
 
-from bigchem.tasks import compute
+current_dir = Path(__file__).resolve().parent
 
-# Instantiate Molecule
-water = qcng.get_molecule("water")
+# Create the molecule
+h2o = Molecule.open(current_dir / "h2o.xyz")
 
-# Create AtomicInput
-my_input = AtomicInput(
-    molecule=water,
+# Define the program input
+prog_input = ProgramInput(
+    molecule=h2o,
+    calctype=CalcType.energy,
     model={"method": "b3lyp", "basis": "6-31g"},
-    driver="energy",
-    keywords={"closed": True, "restricted": True, "purify": False},
-    protocols={"native_files": "all"},  # Request all files from TeraChem Frontend
+    # keywords={"purify": "no"},
 )
 
 # Submit computation to BigChem
-future_result = compute.delay(my_input, "terachem_fe")
+future_output = compute.delay("terachem", prog_input, collect_files=True)
 
 # Check status (optional)
-print(future_result.status)
+print(f"Calculation Status: {future_output.status}")
 
 # Get result from BigChem
-result = future_result.get()
+output = future_output.get()
 
 # Remove result from backend
-future_result.forget()
+future_output.forget()
 
-# See stdout
-pprint(result.stdout)
+### Accessing results ###
+# Stdout from the program
+print(output.stdout)  # or output.pstdout for short
+# Input data used to generate the calculation
+print(output.input_data)
+# Provenance of generated calculation
+print(output.provenance)
 
-# See all natives files
-print(f"Returned files: {result.native_files.keys()}")
+# Check results
+if output.success:
+    print("Energy:", output.results.energy)
+    # The CalcType results will always be available at .return_result
+    print("Energy:", output.return_result)
 
-# Write some returned files to local machine:
-with open("c0", "wb") as f:  # "wb" for binary data
-    f.write(result.native_files["c0"])
+else:  # output.success is False
+    print(output.traceback)  # See why the program failed; output.ptraceback for short
 
-with open("geom.molden", "w") as f:  # "w" for text data
-    f.write(result.native_files["geom.molden"])
 
-with open("tc.out", "w") as f:  # "w" for text data
-    f.write(result.stdout)
+# See all output files
+print(f"Returned files: {output.files.keys()}")
 
-print(result)
-print("Check your directory for a 'c0', 'geom.molden', and 'tc.out' file.")
+# Write the returned files to your local machine
+output.files.write_files()
+print("Check your directory for all of the files produced by TeraChem.")
 
 # This command is not necessary. Sometime a reverse proxy holds open a connection.
 # This closes it gracefully. If instead the following error is raised when the script
@@ -56,4 +61,4 @@ print("Check your directory for a 'c0', 'geom.molden', and 'tc.out' file.")
 # 'ImportError: sys.meta_path is None, Python is likely shutting down'
 # It just means the redis client was holding open a connection and did not get a chance
 # to close it before python exited.
-future_result.backend.client.connection_pool.disconnect()
+# future_output.backend.client.connection_pool.disconnect()