+++ scripts/hacking/__init__.py
@@ -1,6 +1,0 @@
-"""Hacking scripts are random scripts I build to help develop or debug certain features
-
-I keep them around because I may use them in the future to remind me of various ideas,
-but they should not be thought of as "ready" scripts to demonstrate proper usage of BigChem
-in any way.
-"""