+++ CHANGELOG.md
@@ -9,14 +9,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Multiple scripts to `/examples/` directory to show basic utilization of BigChem.
-- [docs/swarm-gpu.md](./docs/swarm-gpus.md) write up covering GPU support in Docker Swarm
-- TeraChem compose and swarm yaml specifications
-- Quickstart documentation to README.md
+- [docs/swarm-gpu.md](./docs/swarm-gpus.md) write up covering GPU support in Docker Swarm.
+- TeraChem compose and swarm yaml specifications.
+- Quickstart documentation to README.md.
+- `qcio` and `qcop` as main data structures and QC program driver packages.
 
 ### Changed
 
 - Dropped `docker-compose` from yaml filenames in `/docker` to make commands less verbose.
-- Modified `docker-compose.yaml` to work for both `compose` and `swarm`
+- Modified `docker-compose.yaml` to work for both `compose` and `swarm`.
+- Removed support for Python 3.7 (reached end of life).
+- Updated worker container from `micromamba:1.3-jammy` -> `micromamba:1.4-jammy`.
+
+### Removed
+
+- Removed support for `QCElemental` and `QCEngine`.
+- `scripts/hacking` directory with old, unused files.
 
 ## [0.4.0] - 2023-02-3